fix types for `UpdateLocationActions`
Type for actions need to be a union instead of an intersection. We want to specify an action as one of and not all of the given methods.

Unfortunately, this causes that the type of arguments in the middleware cannot be narrowed so easily. So we need to handle each constellation of arguments seperately.

Also refactored middleware to do a early return to reduce the nesting a bit.